Fix yarn format git hooks passing for unformatted
Previously, if the staged files are not formatted according to prettier, running
the pre-commit hook would change the file formatting, and then return an exit
code 0, allowing the commit to go through. However, the changes made by prettier
are not staged.

Now, the pre-commit hook will fail, so that only prettier-passed files can be
committed.