Change editorWidth to percentage
This means that the change in width (in pixels) would need to be
converted to a percentage change (in the component) and the string must
be converted, added to the percentage change, then parsed again. A
possble optimization is keeping the number as a float (which would
involve one parse when passing it as a prop). However, I find it
semantically cleaner to keep the `editorWidth` as a string that expresses
a percentage, as compared to an `editorWidthPercentage` that holds a
number.