Scaffold device (#50)
* Rename 'dashboard' -> 'device'

* Add a background image for device

* Add a login dialog, include in device

* Add session state with property token

token is an optional string property of the new slice of the state, session.
token represents an API token returned by the cadet backend. If there is no
token, assume that the user is not logged in.

token is a placeholder property.

* Add action to change the session token

* Connect session.token to Login instead of Device

I'm planning on also requiring log in to access the game. In that case, both
device and game will call the log in dialog. Then, it makes more sense to
connect state.session.token to log in once, rather than twice to device and
game.

* Set up a dummy log in button

* Use SCSS to load background image

Added a variable to define images path as well.