Produce a blank canvas in the Game component
The latest version of pixi was used, because I could only get the
typedefs for 4.7.X. A blank canvas is produced but the story is not
loaded onto it (no attempt was made).

There was a webpack error when I used the older version of pixi, probably due to the discepency between the
typedef and the actual exports.

No verify.