Fix glitchiness of output when updating Repl
I believe that this is due to trying to scroll and update the component
at the same time, causing glitches (such as half the output being
rendered, followed by the other half in 0.2s). I solved this using a
slightly hacky solution -- to wait for a very short period before
scrolling. Ideally, I would like to use a React method that is called
*after* updating.