Make REPL scroll down on I/O change (#67)
* Add componentUpdate function to scroll down

* Fix glitchiness of output when updating Repl

I believe that this is due to trying to scroll and update the component
at the same time, causing glitches (such as half the output being
rendered, followed by the other half in 0.2s). I solved this using a
slightly hacky solution -- to wait for a very short period before
scrolling. Ideally, I would like to use a React method that is called
*after* updating.

* Revert "Fix glitchiness of output when updating Repl"

This reverts commit 6c8bc347ce6d7ec92c258c00a9a0a496f5786ec3.
Using setTimeOut caused an error where, given fast input and output, the
outputs may not be grouped together properly.

* Make div surrounding ReplInput a card

So that it is padded in a similar fashion as ReplOutputs

Resolves #41