Add game (#82)
* Add an empty component for the game

* Copy over source-academy2's story-xml as story

* Produce a blank canvas in the Game component

The latest version of pixi was used, because I could only get the
typedefs for 4.7.X. A blank canvas is produced but the story is not
loaded onto it (no attempt was made).

There was a webpack error when I used the older version of pixi, probably due to the discepency between the
typedef and the actual exports.

No verify.

* Make story javascript runnable

What remains is to pass in the right data to the assets server to get
the story assets.

* Add mock game data, TweenMax CDN

* Add pixi libraries as local scripts

Couldn't find any node packages that fit the version requirements.

* Add scss for the game

* Remove pixi <script> tags

* Add es6 imports for PIXI in story/

Also moved the pixi.js module from devDependencies to dependencies

* Change exports to es6, got Loading... screen

* Fix implicit declarations, got hallway to render

* Fix crash due to pixi-filter's GlowFliters

The GlowFilter constructor from v3 took 7 arguments, but the v4 variant only
takes 5. Removed the first two (height, width). This fixes the runtime error,
but causes the animation for GlowFilter to bug out.

- When an animated object implementing the GlowFilter filter is hovered over,
  (and presumably, the GlowFilter is applied), the animated object will
  'teleport' to the top-left.

* Fix animation obj with glow not rendering

* Fix dialog text causing runtime error

The story's dialog-manager was depending on an undocumented method wordWrap from
version 3 of pixi, that was not ported over to version 4. Fixed by extending
v4's PIXI.text to include this method.

* Make game canvas resize relative to browser

* Relocate pixi story stuff under components

* Remove 1/3 jquery calls with react element ref

* Resolve pixi deprecation warnings

* Make game state

* Fix game not working on second render onwards

* Update redirects in the game

1. Update URLs within source academy
2. Make redirects react-friendly

* Fix typo in comment