Move Game button into Academy
In order to fix a problem with the `object-contain` property for Game, I
added a display:flex property to Academy to allow the Game to take up as
much width as possible (`flex 1 1`) while allowing the academy navbar to
stay in its initial height. To be verbose, one could set flex:initial in
the navbar as well.