Change UX structure (#85)
* Move button to desired structure

This is purely a visual change. The actual structure is to be changed

* Add fixed snapshot

* Fix links from device --> academy

* Move Announcements out of device

* Add username display in place of 'Profile'

* Change all instances of device -> academy

* Change background image name

* Change Device styles to Academy

* Move display stuff from academy to ContentDisplay

Addresses #87

* Move ContentDisplay styles to a new stylesheet

Called _commons.scss

* Remove all-caps for main NavigationBar

* Fix alignment for mission card due date

* Move Game button into Academy

In order to fix a problem with the `object-contain` property for Game, I
added a display:flex property to Academy to allow the Game to take up as
much width as possible (`flex 1 1`) while allowing the academy navbar to
stay in its initial height. To be verbose, one could set flex:initial in
the navbar as well.

* Move Game files into academy

- Changed tslint exception directory
- Changed relative utils dependency directory

* Hide profile button if not logged in

* Remove Game button

* Change url /announcements -> /news

* Fix admin navbar not being thin

Resolves #81, resolves #87