Add a divider to make a margin for the right side
when the width is < 5%. The left margins in the right-parent children
(in `pt-card`s) were removed, and the new div editorDividerDiv acts as
the "margin" when the editor resizable is snapped to the left (0%). When
not snapped, there is a right margin on the whole left-parent to space it
from the right-parent.