Create resizable workspace side content (#89)
* Rename workspace editor/repl parents

* Add dummy component for side content

* Add resizable to side-content

* Fix ability to drag resizable boundary offscreen

* Fix REPL not taking up the rest of the screen

* Fix REPL not flush with editor top

* Change the default side-content text

* Make height of sideContent persist on remount

* Use bounds props to limit max height

Instead of implementing my own messy solution. (A lesson in reading
documentation)

* Make sideContent bottom flush with editor bottom

* Add snapping for Editor's parent Resizable

- Renamed dividerDiv to sideDividerDiv for verbosity

* Add a divider to make a margin for the right side

when the width is < 5%. The left margins in the right-parent children
(in `pt-card`s) were removed, and the new div editorDividerDiv acts as
the "margin" when the editor resizable is snapped to the left (0%). When
not snapped, there is a right margin on the whole left-parent to space it
from the right-parent.