Make NavigationBars Responsive (#93)
* Abstract Navbar style

Into primary and secondary navbars. With this, Admin also shares the
same look as the academy navbar.

REBASE: Used the extension implemented in Change-UX.

* Make Application NavBar responsive

- Added a dependency: flexboxgrid-helpers (for the hidden classes, which
were not added to the original flexboxgrid due to no maintainence. We
could attempt changing our flexboxgrid dependency to flexboxgrid2, which
is a fork of the original.)
- Added a css class to make NavbarDivider thinner on small displays.
- Icon names are hidden on xs devices (having the chrome window half
size, or using a phone)

REBASE: Used the undefined checking in Change-UX

* Fix margin on navbar buttons

Added some documentation for the navbar scss as well.

* Make academy navbar responsive

REBASE: Removed game button

* Make admin navbar responsive

* Fix semantic merge conflicts

- Fix navbargroup height issue
- Use specificity instead of extention for primary & secondary/thin navbar (no need to override, common styles in NavigationBar)
- Add responsiveness for username text

Resolves #71, resolves #76