Standardize parameters for workspace
Now, we can pass in only a value and an optional library number.
This allows the Playground to handle parsing of the library program. If
there is a program passed in via the hash, the parsed program will be
passed. If not, the parsed program will be undefind, causing the redux
store value to be passed (using the `||` operator).