Enable "display" standard library function (#139)
* Add generic type for Context definition

* Rename external to externalContext

* Add typing for createContext & createEmptyContext

* Add externalContext parameter to state

Any instance of createContext() has been given the appropriate
parameter. As a result, the context has a property that corresponds to
the location it is in (externalContext === 'assessment' or 'playground')

* Add location parameter to display and timed (WIP)

timed does not work as of yet, as the initial implementation does not
work in the first place.

* Add working stdlib.timed function

- This does not care about context, and runs the same as map or filter.
I think this should be changed in the future. Unfortunately, I had
little reference as all implementations of `timed` failed when I tried
them (sourceacademy2).

- To make this run, I disabled redux-persist. I shall add what happened
into the tracked issue. To get this to run, checkout createStore.ts and
index.tsx from a commit prior to adding the cache (from master).

- I will move timed back into the "infinity" chapter after getting it to
work.

- Logging does not work as of now.

* Fix logs not showing up

We should find a way to type the action payloads to catch erros like
these (where we are forced to pass things as untyped objects)

* Format files

Missed out a todo there (but its coming up in the next commit)

* Move timed back to orignal place

* Format one whitespace character

* Fix problem with reducers import