Scaffold Grading functionality (#151)
* Add react-router props for Grading/index

* Add param checks and component render call

The call for GradingWorkspaceContainer is added, but the
container/component is not defined yet.

* Add container

* Add scaffolded component

Many todos.

* Format components

* Move GradingOverview to gradingShape.ts

* Add types for Grading and GradingQuestion

* Fix a reference and type error

* Fix interface error

The problem was extending both IMCQQuestion and IProgramming question,
that had clashing properties `type`. Solved this by extending a base
class.

* Fix exports for GradingWorkspace component

* Fix type error for GradingQuestion instance

Also added a call to actions (to be defined)

* Fix references for grading/index

* Add action and action creator

* Fix reference errors in mocks and reducers

* Add mock gradingAPI

* Fix fetching of GradingOverviews

* Add mock Grading and fetch

* Add saga for fetching mock grading

* Add action for updating grading

Also fix some saga errors

* Add slice of state for grading

* Add reducer and extra parameter

This is because Grading is not specified to have an id as of now.

* Add regex for redirect

* Fix typo at reducer

* Fix bug in reducer

* Use AssessmentWorkspace css

They have similar CSS traits. I will consider abstracting this in
another PR, together with the components itself, that are pretty bulky.

* Add a scaffold grading tab

* Format files

* Simplify logic of rendering loading/grid

* Simplify logic for AssessmentWorkspace

* Format Grading/index component

* Use constant value for mock access token