Grading comments (#162)
* Added react-md -- a Markdown editor

- ~400 stars
- In Typescript
- Vertical layout for the slightly narrow side-content section
- In actual Markdown

Only caveats are that it is not WYSIWYG and does not have a markdown to html converter.

* Add draft-js dependency

The main library react-mde runs on.

* Add types for draft-js

* Add showdown dependency

To convert markdown to HTML and vice-versa

* Add awesomefonts dependency

I'll remove this to use blueprint icons instead, later on.

* Add typedefs for showdown

* Add GradingEditor component

* Add react-mde styles into index.scss

Or else, the styles do not work.

* Fix overflow for react-mde in side-content card

* Fix scrolling problems using css

* Add coloring styles

* Add dispatch and state props

Trying to use componentDidMount and componentWillUnmount to handle
updates.

* Add container

* Add action and reducer

* Fix compiler errors

Alphabetisizing, typos.

* Remove undefined union for gradingCommentsValue

The `markdown` value in `mdeState` being undefined causes a runtime
error. So it is easier to just deal with a definite string (since
undefined is not allowed anyway). The typedef for the value allows
undefined though.

* Scaffold link component to render for grid

The problem is that the current <a href> implementation is a hard
redirect, causing a change in URL (and loss of all state). I'm trying to
get a NavLink

* Format and de-style NavLink

* Use undefined check

* Force non-null for markdown value

Since it will be set by the default state,
it is never null.

* Add componentWillMount for grading/index

* Add props for updating ids

* Add action

* Fix buggy componentWillMount functions

Checks only for submissionId, if it is present then there should be a
questionId as well.

* Add defaultComments value

Used to reset the workspace as well.

* Format files

* Add NumericInput for GradingEditor

* Add NumericInput and Button

* Add local state property to track XP given

* Add update for grading XP

* Add action for saving grading input

* Add saga for saving grading

* Format files

* Use blueprintjs icons instead of fontawesome

* Fix save button wrap issue

* Format files

* Remove blank line

* Add some comments

* Format comments