Changed ADSR function to have a linear decay
The previous version of ADSR (Attack, decay, sustain, release) uses an exponential decay function. This may cause issues because the tail end of the decay does not line up with the next part of the sound, which can cause glitches. Thus a linear function is used instead, because the tail end of the decay will match up exactly. There is the added benefit of the function being easier to read and code.