Refactor list visualizer TreeNode
Instead of TreeNode having:
- isFunction
- left, right
- data1, data2

Now, a tree node is either a PairTreeNode, FunctionTreeNode
or DataTreeNode. left, right has the following type:
TreeNode | number | null

When it is a number, it refers to the id of the node to point
to (previously leftid/rightid). When it is a TreeNode, it is
either a DrawableTreeNode or a DataTreeNode. A DrawableTreeNode
is either a PairTreeNode (representing a pair) or a
FunctionTreeNode (representing a function object). A DataTreeNode
represents leaf nodes containing primitive values/array values.