Fix sublanguage rendering logic and update types
Turns out the chapter list rendered has been used wrongly this whole
time, resulting in an inability to filter. After fixing, a lot of the
conditional rendering can be simplified. This is due to an additional
attribute of SALanguage that was created: the main language that a
sublanguage belongs to.