Fix dependency errors
For some reason the types in the lockfile was still stuck to React 17
post-merge. In this commit we reset the lockfile to match `master` and
reran `yarn install` to update the dependencies to match what is
required for this PR. We also migrated one dependency that should have
been a devDependency instead.