Use component state over env state for usingSubst
Previously, whether or not the `Output` component needs to interpret the
incoming output as something for the stepper was dictated by the
environment state. Since the currently showing tab depends on the
component, as opposed to the environment setting, this resulted in a
runtime error when the currently selected tab is not the stepper but the
output tries to interpret the stepper output as a normal output.